[::ZeroMQ] move limit counter to correct level
This has been broken since the move to ZMQ synchronization instead of
mutexes, etc. When the counter was at the top of the poll loop, it was
counting all messages, including those passing through the ZMQ::PAIR
used for main -> thread communication. This caused all manner of funny
business in tests.