Add support for prefixing console output.
This required a surprising amount of change to avoid doing gross hacks. run_readers
now takes an optional argument that is the source actor so the prefix can optionally
be extracted and added to the console output.

On the upside, any proc with an arity of exactly 2 will be passed the actor and can
use it for nefarious purposes.

Also added yard docs and marked some things private.

All Hastur integration tests are still passing.