Revamp IO handling. Again. Add some more YARD. Fix bugs.
Nodule::Stdio just wasn't a practical idea. select() across threads reading from
a single fd was a silly idea and is gone. LineIO is much less complicated and handles
the thread setup so it's not embedded in Process, which was the original reason for Stdio.

Rewrote bits of Process that used Stdio and added a bunch of methods for accessing stdout / stderr
without having to give access to the Nodule instances it has internally.

Updated tests.