fix race condition in process
The background threads weren't coming up fast enough to get all of the output
before a caller could lock the foreground thread up in waitpid. Moved the
start of the background threads before spawn() (which will cause a context switch) and
adjusted LineIO to not return from its run() until the thread is known to be up
and running.